requirements.txt
numpy
pandas
pillow
matplotlib
scikit-image